% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milonga.R
\name{getCumulConvergeStat}
\alias{getCumulConvergeStat}
\title{getCumulConvergeStat}
\usage{
getCumulConvergeStat(res.arr)
}
\arguments{
\item{res.arr}{results object}
}
\description{
getCumulConvergeStat
}
