% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milonga.R
\name{getPatternPrior}
\alias{getPatternPrior}
\title{getPatternPrior}
\usage{
getPatternPrior(pattern, fprior.list)
}
\arguments{
\item{pattern}{pattern matrix}

\item{fprior.list}{list of strings match to the pattern}
}
\description{
getPatternPrior
}
