% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milonga.R
\name{monitorConvergence}
\alias{monitorConvergence}
\title{monitorConvergence}
\usage{
monitorConvergence(sim.list, FUN = monitorStats.default, ...)
}
\arguments{
\item{sim.list}{list of simulation ouput}

\item{FUN}{function}

\item{...}{other parameters}
}
\description{
monitorConvergence
}
