% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milonga.R
\name{nicePrintList}
\alias{nicePrintList}
\title{nicePrintList}
\usage{
nicePrintList(l, col1, col2)
}
\arguments{
\item{l}{length}

\item{col1}{column1}

\item{col2}{column2}
}
\description{
nicePrintList
}
