% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milonga.R
\name{plotConvergence}
\alias{plotConvergence}
\title{plotConvergence}
\usage{
plotConvergence(statdf, outprefix, plot.var = "sqrtR")
}
\arguments{
\item{statdf}{statistics data frame}

\item{outprefix}{prefix for output}

\item{plot.var}{variables to plot}
}
\description{
plotConvergence
}
