% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milonga.R
\name{runMITest}
\alias{runMITest}
\title{runMITest}
\usage{
runMITest(sim.last, alpha = 0.05, null.mean = 0,
  FUN = chainMIVars.default, ...)
}
\arguments{
\item{alpha}{type I error}

\item{null.mean}{mean under NULL}

\item{FUN}{function}

\item{...}{pass to other functions}

\item{sim.list}{list of simulation output}
}
\description{
runMITest
}
