% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milonga.R
\name{milonga}
\alias{milonga}
\title{Multiple imputation for repeated binary outcomes}
\usage{
milonga(data, cols = NULL, nprior = 100, fprior.list = NULL,
  burnin = NULL, rep = 10)
}
\arguments{
\item{data}{Data frame or matrix of multivariate binary outcome with missing values to impute, NA denotes missing values}

\item{cols}{NULL or a character or integer vector giving the columns for imputation}

\item{nprior}{integer}

\item{fprior.list}{a list of prior frequencies. See "Details"}

\item{burnin}{number of burnin iterations, NULL or positive integer}

\item{rep}{number of iterations for imputation, default is 10}
}
\value{
list of imputed data frames
}
\description{
Multiple imputation for repeated binary outcomes
}
\examples{
data(polio)
milonga(polio, cols=2:7)
}
